requirements.txt
-i https://pypi.org/simple
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.3
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
